Feat: 카테고리별 태그 색상 구분 기능 추가
입력되는 내역들의 카테고리에 따라서 태그 색깔이 다르게 나타나도록 했다.
style.css 파일에 클래스 이름별로 미리 각 색상들을 background-color로
지정해놓고 렌더링 할때 카테고리에 맞는 클래스를 추가하는 방식으로
구현하였다.

그리고 beColorList 라는 객체를 constants.js 파일에서 import 해오는 방식을
이용했다.
