+++ index.html
@@ -214,12 +214,12 @@
 
 		<main class="monthly-list">
 			<header class="info-filter">
-				<div>전체 내역 13건</div>
+				<div>전체 내역 0건</div>
 				<div>
 					<input class="filter-checkbox income" type="checkbox" checked>
-					<label for="">수입 500,000</label>
+					<label for="">수입 0</label>
 					<input class="filter-checkbox expenditure" type="checkbox" checked>
-					<label for="">지출 27,000</label>
+					<label for="">지출 0</label>
 				</div>
 			</header>
 			<section class="main-list">
