+++ src/css/main.css
@@ -75,7 +75,6 @@
 	padding-top: 4px;
 	height: 32px;
 	text-align: center;
-	background-color: var(--color-category-mint);
 	color: var(--color-white);
 	font: var(--font-bold-medium);
 	flex-basis: 92px;
