+++ src/js/constants/constants.js
@@ -1,0 +1,27 @@
+export const monthNameList = {
+	1: 'January',
+	2: 'Feburary',
+	3: 'March',
+	4: 'April',
+	5: 'May',
+	6: 'June',
+	7: 'July',
+	8: 'August',
+	9: 'September',
+	10: 'October',
+	11: 'November',
+	12: 'December',
+};
+
+export const bcColorList = {
+	생활: 'bc-navy',
+	식비: 'bc-blue',
+	교통: 'bc-mint',
+	'쇼핑/뷰티': 'bc-cyan',
+	'의료/건강': 'bc-sky',
+	'문화/여가': 'bc-pink',
+	미분류: 'bc-purple',
+	현금: 'bc-green',
+	신용카드: 'bc-yellow',
+	계좌: 'bc-orange',
+};
