+++ src/js/header/prevBtnHandler.js
@@ -1,4 +1,4 @@
-import { monthNameList } from '../constants/monthNameList.js';
+import { monthNameList } from '../constants/constants.js';
 
 export function prevBtnHandler() {
 	const currentYearEl = document.querySelector('div.current-month-year p:nth-of-type(1)');
